Update ubuntu version to Focal Fossa and fix minor typo
Whenever the ubuntu version is mentioned, it was changed from `latest`
to `focal`.
This change was done because of a found error while building the nginx
server from source: it raisen an error because at this time (2023-04-19)
no packages for `openssl1.1` have been published for the jammy-jellyfish
ubuntu release.

The easiest fix was to lock the ubuntu version to `focal-fossa`, which
has the package ready for install from `apt`.

Also, at ln 379 was fixed a double bracket typo.